d_a_obj_buoyflag: Improve calc_pos (USA partial)

+++ src/d/actor/d_a_obj_buoyflag.cpp
@@ -450,7 +450,7 @@ void daObjBuoyflag::Packet_c::calc_pos_spring_near(const cXyz* pos, const cXyz* 
 
 /* 000015FC-00001BC0       .text calc_pos__Q213daObjBuoyflag8Packet_cFPQ213daObjBuoyflag5Act_c */
 void daObjBuoyflag::Packet_c::calc_pos(Act_c* actor) {
-    // USA: remaining differences are instruction scheduling and register allocation.
+    // USA: remaining difference is the placement of 32768.0f * distance in calc_pos_wave.
     DrawVtx_c* draw = &mDraw[mBuffer];
     DrawVtx_c* prev = &mDraw[mBuffer ^ 1];
     calc_wind_base(actor);
@@ -462,7 +462,8 @@ void daObjBuoyflag::Packet_c::calc_pos(Act_c* actor) {
             calc_pos_gravity(y, x);
             calc_pos_wave(y, x);
             calc_pos_spd(y, x);
-            draw->pos[y][x] = prev->pos[y][x] + mMove.speed[y][x];
+            // Fakematch: the unsigned index keeps the speed address from being reused from calc_pos_spd.
+            draw->pos[y][x] = prev->pos[y][x] + mMove.speed[y][(u32)x];
         }
     }
 }
